Change timeout behavior
Instead of setting the timeout on the whole XHR request, a timeout is triggered when no transfer progress is being made for the specified time.
E.g. with a slow upload speed, a 5MB chunk could take several minutes. With the old timeout behavior this would mean that the timeout would have to be set to e.g. 10 minutes to ensure successful uploads. With the new behavior that is no problem. It is only detected as a timeout if the transfer actually stalls longer than the timeout is.